\name{check.data.structure}
\alias{check.data.structure}
\title{Issues a stop if the data does not conform with the expected structure}
\usage{
  check.data.structure(dat)
}
\arguments{
  \item{dat}{data for use in CDT}
}
\value{
  NULL
}
\description{
  Issues a stop if the data does not conform with the
  expected structure
}
\author{
  Andrew Azman
}

