% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.fit.R
\name{dgammaOff1}
\alias{dgammaOff1}
\title{Function that calculates dgamma with a offset of 1 (i.e., 1 is equivlaent to 0)}
\usage{
dgammaOff1(x, replace0 = FALSE, ...)
}
\arguments{
\item{x}{value to calculate dgamma at}

\item{replace0}{should we replace 0 with episilon}

\item{...}{other parameters to dgamma}
}
\value{
dgamma offset
}
\description{
Function that calculates dgamma with a offset of 1 (i.e., 1 is equivlaent to 0)
}
