\name{diclik}
\alias{diclik}
\title{calculates the DIC likelihood by integration}
\usage{
  diclik(par1, par2, EL, ER, SL, SR, dist)
}
\arguments{
  \item{par1}{}

  \item{par2}{}

  \item{EL}{}

  \item{ER}{}

  \item{SL}{}

  \item{SR}{}

  \item{dist}{}
}
\description{
  calculates the DIC likelihood by integration
}

