\name{diclik2.helper1}
\alias{diclik2.helper1}
\title{likelihood functions for diclik2}
\usage{
  diclik2.helper1(x, SL, SR, par1, par2, dist)
}
\arguments{
  \item{x}{}

  \item{SL}{}

  \item{SR}{}

  \item{par1}{}

  \item{par2}{}

  \item{dist}{}
}
\description{
  likelihood functions for diclik2
}

