\name{diclik2}
\alias{diclik2}
\title{this dic likelihood is designed for data that has overlapping intervals}
\usage{
  diclik2(par1, par2, EL, ER, SL, SR, dist)
}
\arguments{
  \item{par1}{}

  \item{par2}{}

  \item{EL}{}

  \item{ER}{}

  \item{SL}{}

  \item{SR}{}

  \item{dist}{}
}
\description{
  this dic likelihood is designed for data that has
  overlapping intervals
}

