\name{dist.optim.transform}
\alias{dist.optim.transform}
\title{Transforms parameters of a specific distriution for unbounded optimization}
\usage{
  dist.optim.transform(dist, pars)
}
\arguments{
  \item{dist}{string representing distirbution}

  \item{pars}{vector of parameters}
}
\value{
  vector of transformed parameters
}
\description{
  Transforms parameters of a specific distriution for
  unbounded optimization
}

