\name{dist.optim.untransform}
\alias{dist.optim.untransform}
\title{Untransforms parameters before entering likelihood}
\usage{
  dist.optim.untransform(dist, pars)
}
\arguments{
  \item{dist}{}

  \item{pars}{}
}
\value{
  vector of untransformed parameters
}
\description{
  Untransforms parameters before entering likelihood
}

