\name{fw1}
\alias{fw1}
\title{the functions coded here are taken directly from the
doubly interval censored likelihood notes.}
\usage{
  fw1(t, EL, ER, SL, SR, par1, par2, dist)
}
\arguments{
  \item{t}{}

  \item{EL}{}

  \item{ER}{}

  \item{SL}{}

  \item{SR}{}

  \item{par1}{}

  \item{par2}{}

  \item{dist}{}
}
\description{
  the functions coded here are taken directly from the
  doubly interval censored likelihood notes.
}

