\name{logLik}
\alias{logLik}
\alias{logLik,cd.fit-method}
\title{Get the log-likelihood value of a \code{cd.fit} or \code{cd.fit.mcmc} object}
\arguments{
  \item{object}{A \code{cd.fit} or \code{cd.fit.mcmc}
  object.}
}
\value{
  log-likelihood value
}
\description{
  Get the log-likelihood value of a \code{cd.fit} or
  \code{cd.fit.mcmc} object
}

