\name{plot}
\alias{plot}
\alias{plot,cd.fit-method}
\title{Plots the estimated survival function with an option to plot the posterior draws or bootstraps behind the fit}
\arguments{
  \item{x}{\code{cd.fit} or \code{cd.fit.mcmc} object}

  \item{col.main}{color for plotting the main estimate}

  \item{col.samps}{color for the samples (should include
  some alpha transparency)}

  \item{plot.n.samps}{how many posterior or boostrap
  samples do you want to plot?}

  \item{add}{will add to exisiting plot}

  \item{...}{other options to pass to plot}
}
\description{
  Plots the estimated survival function with an option to
  plot the posterior draws or bootstraps behind the fit
}

