\name{plot}
\alias{plot}
\title{For now this is going to plot the estimated survival
function and if bootstrap or mcmc samples are present, it will plot these samples with alpha}
\arguments{
  \item{x}{}

  \item{y}{}

  \item{col.main}{- color for plotting the main estimate}

  \item{col.samps}{- color for the samples (should include
  some alpha transparency)}

  \item{plot.n.samps}{- how many posterior or boostrap
  samples do you want to plot?}

  \item{...}{- other options to pass to plot}
}
\description{
  For now this is going to plot the estimated survival
  function and if bootstrap or mcmc samples are present, it
  will plot these samples with alpha
}

