\name{get.obs.type}
\alias{get.obs.type}
\title{Tries to guess the observation types (SIC, DIC, or exact).}
\usage{
  get.obs.type(dat)
}
\arguments{
  \item{dat}{a matrix of data, similar to what needs to be
  passed to \code{dic.fit()}.}
}
\value{
  vector of guessed types
}
\description{
  Tries to guess the observation types (SIC, DIC, or
  exact).
}

