% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4stuff.R
\name{logLik,cd.fit-method}
\alias{logLik,cd.fit-method}
\alias{logLik}
\title{Get the log-likelihood value of a \code{cd.fit} or \code{cd.fit.mcmc} object}
\usage{
\S4method{logLik}{cd.fit}(object)
}
\arguments{
\item{object}{A \code{cd.fit} or \code{cd.fit.mcmc} object.}
}
\value{
log-likelihood value
}
\description{
Get the log-likelihood value of a \code{cd.fit} or \code{cd.fit.mcmc} object
}
